MNT #256 relocate helpers, remove unittest legacy
